compart_id 0 does mean something so cannot be used as the unassign value, changing it to -1

+++ includes/rtld_linkmap_scan.h
@@ -37,7 +37,7 @@
 #define __CHERI_PURE_CAPABILITY__
 
 typedef struct struct_compart_data_t {
-	uint16_t id;
+	int id;
 	char *path;
 } compart_data_t;
